pyproject.toml
[project]
name = "mogle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "isort>=6.0.1",
    "jose>=1.0.0",
    "jwt>=1.3.1",
    "mypy>=1.16.1",
    "passlib>=1.7.4",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.12.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
